backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
openai
psycopg2-binary  # if using Postgres later
